refactor(frontend): rename and relocate test utilities
- rename using camel-case for consistency
- move test setup file to utils
